requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
mysqlclient==2.2.4
